% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_content}
\alias{clean_content}
\title{Clean emails}
\usage{
clean_content(content)
}
\arguments{
\item{content}{list of content as returned by \code{\link{get_content}}.}
}
\description{
Clean emails by removing \emph{useless} lines.
}
\details{
Example of line removed
\code{UNCLASSIFIED U.S. Department of State Case No. F-2014-20439 Doc No. C05765911 Date: 08/31/2015}
look at the source code for more details \code{clean_content}.
}
\examples{
\dontrun{
hrc_emails <- load_emails(emails_bengh) # load emails
cont <- get_content(hrc_emails)
cont <- clean_content(hrc_emails)
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}
