% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_emails}
\alias{clean_emails}
\title{Clean emails}
\usage{
clean_emails(emails)
}
\arguments{
\item{emails}{list of emails as returned by \code{\link{load_emails}}.}
}
\description{
Clean emails by removing \emph{useless} lines.
}
\details{
Example of line removed
\code{UNCLASSIFIED U.S. Department of State Case No. F-2014-20439 Doc No. C05765911 Date: 08/31/2015}
look at the source code for more details \code{clean_emails}.
}
\examples{
\dontrun{
emails <- load_emails(emails_bengh) # load emails
emails <- clean_emails(emails)
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}
