% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emails.R
\name{download_emails}
\alias{download_emails}
\title{Download emails}
\usage{
download_emails(release, save.dir = getwd())
}
\arguments{
\item{release}{Name of the batch of release of emails; see details.}

\item{save.dir}{Directory where to save the downloaded emails, defaults to
\code{getwd()}}
}
\value{
Returns full path to downloaded zip or tar.
}
\description{
Download emails manually
}
\details{
Below are the valid values for \code{release}; follows the
\href{http://graphics.wsj.com/hillary-clinton-email-documents/}{WSJ} naming
convention.
\itemize{
\item Benghazi
\item June
\item July
\item August
\item September
\item October
\item November
\item January 7
\item January 29
\item February 19
\item february 29
\item December
\item Non-disclosure
}
}
\seealso{
\code{\link{get_xpdf}}, \code{\link{extract_emails}}
}
\author{
John Coene \email{jcoenep@gmail.com}
}
