% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{edges_emails}
\alias{edges_emails}
\title{Network the treacherous}
\usage{
edges_emails(emails = emails, ...)
}
\arguments{
\item{emails}{Data frame of emails as returned by \code{\link{search_emails}},
defaults to \code{emails} (see \code{\link{emails}})}

\item{...}{any additional column to keep as meta-data}
}
\description{
Builds edge list from emails using \code{from} and \code{to},
edge source and target respectively.
}
\examples{
\dontrun{
emails <- search_emails()

edges <- edges_emails(emails)
}

}
\seealso{
\code{\link{search_emails}}
}
\author{
John Coene \email{jcoenep@gmail.com}
}
