% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_address}
\alias{extract_address}
\title{Extract email addresses}
\usage{
extract_address(emails)
}
\arguments{
\item{emails}{list of email contents, as returned by \code{\link{load_emails}}}
}
\value{
named \code{list} of dates where names are emails/files name.
}
\description{
Exctract email addresses from emails
}
\examples{
\dontrun{
# get xpdf extractor
ext <- get_xpdf()

# create
dir.create("emails")

# get emails released in august
emails_aug <- get_emails(release = "August", save.dir = "./emails",
                     extractor = ext)

# use manually downloaded extractor
# ext <- "C:/xpdfbin-win-3.04/bin64/pdftotext.exe"

# get emails related to Benghazi released in December
emails_bengh <- get_emails(release = "Benghazi", extractor = ext,
                           save.dir = "./emails")

contents <- load_emails(emails_bengh)

addresses <- extract_address(contents)
}

}
\seealso{
\code{\link{load_emails}}
}
\author{
John Coene \email{jcoenep@gmail.com}
}
