% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emails.R
\name{extract_contents}
\alias{extract_contents}
\title{Extract emails contents}
\usage{
extract_contents(emails, save.dir = getwd(), extractor)
}
\arguments{
\item{emails}{directory of folder including all pdf emails.}

\item{save.dir}{destination folder where \code{txt} files will be extracted to.}

\item{extractor}{path to extractor, see \code{\link{get_xpdf}}.}
}
\description{
Extract content of manually downloaded emails.
}
\examples{
\dontrun{
# emails manually downloaded from:
# http://graphics.wsj.com/hillary-clinton-email-documents/zips/HRC_Email_296.zip

# create directory to store extracted contents
dir.create("emails_txt")

ext <- get_xpdf()

extarct_contents(emails = "HRC_Email_296", dest = "./emails_txt", extractor = ext)
}

}
\seealso{
\code{\link{get_xpdf}}
}
\author{
John Coene \email{jcoenep@gmail.com}
}
