% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_com}
\alias{get_com}
\alias{get_com.rodham}
\title{get_com: get emails sender's and receiver's name}
\usage{
get_com(emails)

\method{get_com}{rodham}(emails)
}
\arguments{
\item{emails}{list of email contents, as returned by \code{\link{load_emails}}}
}
\value{
data.frame of names and document id.
}
\description{
Get emails communication
}
\details{
Get the sender's and receiver's name.
}
\examples{
\dontrun{
emails <- load_emails("emails")
com <- get_com(emails)
}

}
