% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_content}
\alias{get_content}
\alias{get_content.rodham}
\title{get_contents: get original emails senders and receivers' name}
\usage{
get_content(emails)

\method{get_content}{rodham}(emails)
}
\arguments{
\item{emails}{list of email contents, as returned by \code{\link{load_emails}}}
}
\value{
named list (document id) of email contents.
}
\description{
Get emails original communication
}
\details{
Get the senders and receivers' name of original email.
}
\examples{
\dontrun{
emails <- load_emails("emails")
contents <- get_content(emails)
}

}
