% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_date}
\alias{get_date}
\alias{get_date.rodham}
\title{get_date: get emails date}
\usage{
get_date(emails)

\method{get_date}{rodham}(emails)
}
\arguments{
\item{emails}{list of email contents, as returned by \code{\link{load_emails}}}
}
\value{
data.frame of dates and document id.
}
\description{
Get emails date
}
\details{
Get the date on which the emails is received.
}
\examples{
\dontrun{
emails <- load_emails("emails")
dates <- get_date(emails)
}

}
