% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_from}
\alias{get_from}
\alias{get_from.rodham}
\title{get_from: get emails sender's name}
\usage{
get_from(emails)

\method{get_from}{rodham}(emails)
}
\arguments{
\item{emails}{list of email contents, as returned by \code{\link{load_emails}}}
}
\value{
data.frame of names and document id.
}
\description{
Get emails sender
}
\details{
Get the sender's name.
}
\examples{
\dontrun{
emails <- load_emails("emails")
dates <- get_date(emails)
}

}
