% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_id}
\alias{get_id}
\alias{get_id.rodham}
\title{get_id: get emails subjects}
\usage{
get_id(emails)

\method{get_id}{rodham}(emails)
}
\arguments{
\item{emails}{list of email contents, as returned by \code{\link{load_emails}}}
}
\value{
vector of emails' document ids.
}
\description{
Get emails ID
}
\details{
Get emails' document id
}
\examples{
\dontrun{
emails <- load_emails("emails")
docids <- get_id(emails)
}

}
