% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_interest}
\alias{get_interest}
\alias{get_interest.rodham}
\title{get_interest: get emails subjects}
\usage{
get_interest(emails)

\method{get_interest}{rodham}(emails)
}
\arguments{
\item{emails}{list of email contents, as returned by \code{\link{load_emails}}}
}
\value{
data.frame of emails' interest and document id.
}
\description{
Get emails interest
}
\details{
Get emails' subjects
}
\examples{
\dontrun{
emails <- load_emails("emails")
subjects <- get_interest(emails)
}

}
