% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_or}
\alias{get_or}
\alias{get_or.rodham}
\title{get_or: get emails senders and receivers' name}
\usage{
get_or(emails)

\method{get_or}{rodham}(emails)
}
\arguments{
\item{emails}{list of email contents, as returned by \code{\link{load_emails}}}
}
\value{
data.frame of names and document id.
}
\description{
Get emails original communication
}
\details{
Get the senders and receivers' name of original email.
}
\examples{
\dontrun{
emails <- load_emails("emails")
original <- get_or(emails)
}

}
