% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{get_xpdf}
\alias{get_xpdf}
\title{Get pdf to text extractor (xpdf)}
\usage{
get_xpdf(dest = getwd())
}
\arguments{
\item{dest}{Destination folder defaults to \code{getwd()}}
}
\value{
Returns full path to \code{pdftotext} executable
}
\description{
Downloads and extracts pdf to text extractor, see details.
}
\details{
If the function fails you can download the
extractor manually from \url{http://www.foolabs.com/xpdf/}
then set it manually as shown in examples \code{\link{get_emails}}

Tested on:
\itemize{
  \item{Windows}
  \item{Linux}
}
}
\seealso{
\code{\link{get_emails}}
}
\author{
John Coene \email{jcoenep@gmail.com}
}
