% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrcnames.R
\docType{data}
\name{hrc_names}
\alias{hrc_names}
\title{Hillary Rodham Clinton emails}
\format{A data frame with 912 rows and 2 variables:
\describe{
  \item{originalName}{Original name}
  \item{commonName}{Commonly-used name in emails}
}}
\source{
\url{https://github.com/wsjdata/clinton-email-cruncher/blob/master/HRCEMAIL_names.csv}
}
\usage{
hrc_names
}
\description{
List that pairs sender and recipient names provided by the State Department website
with that person's commonly-used name.
}
\details{
For example, \code{HRC} becomes \code{Hillary Clinton}.
}
\keyword{datasets}
