% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{nodes_emails}
\alias{nodes_emails}
\title{Build nodes}
\usage{
nodes_emails(emails)
}
\arguments{
\item{emails}{Data frame of emails as returned by \code{\link{search_emails}}}
}
\description{
Build list of nodes and add meta-data.
}
\examples{
\dontrun{
emails <- search_emails()

nodes <- nodes_emails(emails)
}

}
\author{
John Coene \email{john.coene@gmail.com}
}

