% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{rm_comments}
\alias{rm_comments}
\title{Remove Comments}
\usage{
rm_comments(emails)
}
\arguments{
\item{emails}{list of emails as returned by \code{\link{load_emails}}.}
}
\description{
Remove comments
}
\details{
Example of comment removed
\code{UNCLASSIFIED U.S. Department of State Case No. F-2014-20439 Doc No. C05765911 Date: 08/31/2015}
}
\examples{
\dontrun{
emails <- load_emails(emails_bengh) # load emails
emails <- rm_comments(emails)
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}
