% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_emails}
\alias{search_emails}
\title{Search Rodham's emails}
\usage{
search_emails(subject = NULL, to = NULL, from = NULL, start = NULL,
  end = NULL)
}
\arguments{
\item{subject}{Filter by subject, defaults to \code{NULL}(no filter).}

\item{to}{Filter by Receiver, defaults to \code{NULL}(no filter).}

\item{from}{Filter by Sender, defaults to \code{NULL}(no filter).}

\item{start}{Filter by date range, defaults to \code{NULL}(no filter).}

\item{end}{Filter by date range, defaults to \code{NULL}(no filter).}
}
\description{
Search Hillary Rodham Clinton's \emph{personal} emails.
}
\details{
There are a total of 29444 emails ranging from \code{2009-08-14} to
\code{2014-08-13}.
}
\examples{
\dontrun{
emails <- search_emails()
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}

