% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{tidy_emails}
\alias{tidy_emails}
\title{Tidy contents}
\usage{
tidy_emails(content)
}
\arguments{
\item{content}{email content as returned by \code{\link{get_content}}.}
}
\value{
A two-column \code{tibble} with emails document id in one column (\code{emails}) and the email \code{content} in another.
}
\description{
Tidy email contents
}
\examples{
\dontrun{
content <- get_content(content)
content <- clean_content(content)
tidy <- tidy_emails(content)
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}
