% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{get_extractor}
\alias{get_extractor}
\title{Get pdf to text parser (xpdf)}
\usage{
get_extractor(dest = "C:/")
}
\arguments{
\item{dest}{Destination folder defaults to \code{C:/}}
}
\value{
Returns full path to pdftotext.exe
}
\description{
Downloads and extracts pdf to text extractor, see details.
}
\details{
If function fails (sorry - tested on Windows) you can download the extractor manully
from \url{http://www.foolabs.com/xpdf/}
}
\author{
John Coene \email{jcoenep@gmail.com}
}
\seealso{
\code{\link{get_emails}}
}

