% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{get_xpdf}
\alias{get_xpdf}
\title{Get pdf to text extractor (xpdf)}
\usage{
get_xpdf(dest = getwd())
}
\arguments{
\item{dest}{Destination folder defaults to \code{getwd()}}
}
\value{
Returns full path to pdftotext executable
}
\description{
Downloads and extracts pdf to text extractor, see details.
}
\details{
If function fails you can download the
extractor manually from \url{http://www.foolabs.com/xpdf/}
(tested on Windows only)
}
\author{
John Coene \email{jcoenep@gmail.com}
}
\seealso{
\code{\link{get_emails}}
}

