% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodham.R
\docType{package}
\name{rodham}
\alias{rodham}
\alias{rodham-package}
\title{rodham: search Hillary's emails}
\description{
rodham: search Hillary's emails
}
\section{Functions}{

\itemize{
\item \code{\link{search_emails}}
\item \code{\link{edges_emails}}
\item \code{\link{nodes_emails}}
}
}
\examples{
\dontrun{
# fetch emails
emails <- search_emails()

# build graph
edges <- edges_emails()
nodes <- nodes_emails
g <- igraph::graph.data.frame(edges, vertices = nodes)

plot(g)
}

}

